+++ source/Sailfish.TestAdapter/TestSettingsParser/SailfishSettings.cs
@@ -12,4 +12,7 @@ public class SailfishSettings
 
     [JsonPropertyName("SampleSizeOverride")]
     public int? SampleSizeOverride { get; set; }
+
+    [JsonPropertyName("EnableEnvironmentHealthCheck")]
+    public bool? EnableEnvironmentHealthCheck { get; set; }
 }