+++ source/Sailfish/Registration/SailfishModuleRegistrations.cs
@@ -20,6 +20,8 @@ using Sailfish.Presentation;
 using Sailfish.Presentation.Console;
 using Sailfish.Presentation.CsvAndJson;
 using Sailfish.Presentation.Markdown;
+using Sailfish.Diagnostics.Environment;
+
 
 namespace Sailfish.Registration;
 
@@ -111,6 +113,8 @@ internal class SailfishModuleRegistrations : IProvideAdditionalRegistrations
         // Register enhanced SailDiffResultMarkdownConverter with unified formatter
         builder.RegisterType<SailDiffResultMarkdownConverter>().As<ISailDiffResultMarkdownConverter>();
         builder.RegisterType<SailfishExecutionEngine>().As<ISailfishExecutionEngine>().InstancePerDependency();
+
+        builder.RegisterType<EnvironmentHealthReportProvider>().As<IEnvironmentHealthReportProvider>().SingleInstance();
 
         // Register enhanced MarkdownTableConverter with unified formatter
         builder.RegisterType<MarkdownTableConverter>().As<IMarkdownTableConverter>().InstancePerDependency();